fix: resolve mypy errors in deepened modules
orbit_following: wrap numpy scalar returns in float()
ntm_dynamics: replace **dict unpacking with closure for RK4
current_drive: disambiguate float|ndarray returns
detachment_controller: float() wrap for two_point_q_parallel
integrated_scenario: np.asarray() on 6 array-returning helpers

Zero mypy errors outside pre-existing gk_nonlinear/validation.

+++ src/scpn_control/control/detachment_controller.py
@@ -99,7 +99,7 @@ def two_point_q_parallel(T_upstream_eV: float, L_parallel_m: float) -> float:
     """
     if T_upstream_eV <= 0.0 or L_parallel_m <= 0.0:
         return 0.0
-    return _KAPPA_0_SPITZER * T_upstream_eV**3.5 / (7.0 * L_parallel_m)
+    return float(_KAPPA_0_SPITZER * T_upstream_eV**3.5 / (7.0 * L_parallel_m))
 
 
 class DetachmentController:
